+++ src/Jigoshop/Entity/Product/Variable.php
@@ -186,7 +186,9 @@ class Variable extends Product implements Shippable, Saleable
                     $attributes[$attribute->getAttribute()->getId()]['options'][$attribute->getValue()] = $attribute->getAttribute()->getOption($attribute->getValue())->getLabel();
                 } else {
                     foreach($attribute->getAttribute()->getOptions() as $option) {
-                        $attributes[$attribute->getAttribute()->getId()]['options'][$option->getId()] = $option->getLabel();
+                        if(in_array($option->getId(), $attribute->getAttribute()->getValue())){
+                            $attributes[$attribute->getAttribute()->getId()]['options'][$option->getId()] = $option->getLabel();
+                        }
                     }
                 }
             }