+++ src/Jigoshop/Helper/Forms.php
@@ -15,6 +15,7 @@ class Forms
 	protected static $hiddenTemplate = 'forms/hidden';
 	protected static $textareaTemplate = 'forms/textarea';
 	protected static $daterangeTemplate = 'forms/daterange';
+	protected static $userDefinedTemplate = 'forms/userDefined';
 
 	/**
 	 * Returns string for checkboxes if value is checked (value and current are the same).
@@ -96,6 +97,9 @@ class Forms
 			case 'daterange':
 				self::daterange($field);
 				break;
+			case 'user_defined':
+				self::userDefined($field);
+				break;				
 			default :
 				do_action('jigoshop\helper\forms\custom', $type, $field);
 				break;
@@ -118,6 +122,7 @@ class Forms
 	 *   * tip (false) - tip for the tag
 	 *   * hidden (false) - whether to hide element by default
 	 *   * size (12) - default size of the element (Bootstrap column size 12)
+	 * 	 * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -140,6 +145,7 @@ class Forms
 			'tip' => false,
 			'hidden' => false,
 			'size' => 12,
+			'data' => []
         ];
 		$field = wp_parse_args($field, $defaults);
 
@@ -179,6 +185,7 @@ class Forms
 	 *   * min (false) - minimal value of number input
 	 *   * max (false) - maximal value of number input
 	 *   * step (1) - Step of number
+	 *   * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -204,6 +211,7 @@ class Forms
 			'min' => false,
 			'max' => false,
 			'step' => 1,
+			'data' => []
         ];
 
 		$field = wp_parse_args($field, $defaults);
@@ -254,6 +262,7 @@ class Forms
 	 *   * options (array) - available options to select
 	 *   * hidden (false) - whether to hide element by default
 	 *   * size (12) - default size of the element (Bootstrap column size 12)
+	 *   * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -278,6 +287,7 @@ class Forms
 			'hidden' => false,
 			'size' => 12,
             'args' => [],
+            'data' => []
         ];
 		$field = wp_parse_args($field, $defaults);
 
@@ -339,6 +349,7 @@ class Forms
 	 *   * tip (false) - tip for the tag
 	 *   * hidden (false) - whether to hide element by default
 	 *   * size (12) - default size of the element (Bootstrap column size 12)
+	 *   * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -362,6 +373,7 @@ class Forms
 			'tip' => false,
 			'hidden' => false,
 			'size' => 12,
+			'data' => []
         ];
 		$field = wp_parse_args($field, $defaults);
 
@@ -401,6 +413,7 @@ class Forms
 	 *   * tip (false) - tip for the tag
 	 *   * hidden (false) - whether to hide element by default
 	 *   * size (12) - default size of the element (Bootstrap column size 12)
+	 *   * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -422,6 +435,7 @@ class Forms
 			'tip' => false,
 			'hidden' => false,
 			'size' => 12,
+			'data' => []
         ];
 		$field = wp_parse_args($field, $defaults);
 
@@ -457,6 +471,7 @@ class Forms
 	 *   * description (false) - description of the tag
 	 *   * tip (false) - tip for the tag
 	 *   * size (12) - default size of the element (Bootstrap column size 12)
+	 *   * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -479,6 +494,10 @@ class Forms
 			'size' => 12,
 			'startDate' => false,
 			'endDate' => false,
+			'data' => [
+				'from' => [],
+				'to' => []
+			]
         ];
 		$field = wp_parse_args($field, $defaults);
 
@@ -504,6 +523,7 @@ class Forms
 	 *   * name (null) - HTML name for the tag
 	 *   * value (false) - HTML value of the tag
 	 *   * classes (array()) - list of HTML classes for the tag
+	 *   * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -518,6 +538,7 @@ class Forms
 			'name' => null,
 			'value' => false,
 			'classes' => [],
+			'data' => []
         ];
 		$field = wp_parse_args($field, $defaults);
 
@@ -552,6 +573,7 @@ class Forms
 	 *   * tip (false) - tip for the tag
 	 *   * hidden (false) - whether to hide element by default
 	 *   * size (12) - default size of the element (Bootstrap column size 12)
+	 *   * data (array()) - key-value pairs for data attributes
 	 *
 	 * Field's name is required.
 	 *
@@ -574,6 +596,7 @@ class Forms
 			'size' => 11,
 			'startDate' => false,
 			'endDate' => false,
+			'data' => []
         ];
 		$field = wp_parse_args($field, $defaults);
 
@@ -593,6 +616,39 @@ class Forms
 
 		Render::output(static::$constantTemplate, $field);
 	}
+
+	/**
+	 * Outputs already rendered field.
+	 *
+	 * Available parameters (with defaults):
+	 *   * name (null) - name for the tag
+	 *   * label (null) - label for the tag
+	 *   * tip (false) - tip for the tag
+	 *   * display (null) - callback for display content
+	 * 
+	 * @param $field array Field parameters.
+	 *
+	 * @throws \Jigoshop\Exception
+	 */
+	public static function userDefined($field) {
+		$defaults = [
+			'name' => '',
+			'title' => '',
+			'tip' => '',
+			'display' => ''
+		];
+
+		$field = wp_parse_args($field, $defaults);
+		if(isset($field['display']) && is_callable($field['display'])) {
+			$result = call_user_func($field['display'], $field);
+
+			Render::output(static::$userDefinedTemplate, [
+					'title' => isset($field['title'])?$field['title']:'',
+					'tip' => isset($field['tip'])?$field['tip']:'',
+					'display' => $result
+				]);
+		}
+	}	
 
 	public static function printHiddenFields($fields, $exceptions = [])
 	{