+++ templates/forms/constant.php
@@ -10,6 +10,7 @@
  * @var $description string Field description.
  * @var $hidden boolean Whether the field is hidden.
  * @var $size int Size of form widget.
+ * @var $data array Key-value pairs for data attributes.
  */
 ?>
 <div class="form-group <?= $id; ?>_field <?= join(' ', $classes); ?> clearfix<?php $hidden and print ' not-active'; ?>">
@@ -20,7 +21,15 @@
 		<?php endif; ?>
 	</label>
 	<div class="col-sm-<?= $size; ?>">
-		<p class="form-control-static <?= join(' ', $classes); ?>" id="<?= $id; ?>"><?= $value; ?></p>
+		<p class="form-control-static <?= join(' ', $classes); ?>" id="<?= $id; ?>"
+		<?php 
+		if(isset($data) && is_array($data)) {
+			foreach($data as $dataKey => $dataValue) {
+				echo sprintf(' data-%s="%s"', $dataKey, $dataValue);
+			}
+		}
+		?>
+		><?= $value; ?></p>
 		<?php if(!empty($description)): ?>
 			<span class="help-block"><?= $description; ?></span>
 		<?php endif; ?>