+++ templates/shop/cart/mobile.php
@@ -1,8 +1,6 @@
 <?php
 /**
  * @var $cart \Jigoshop\Entity\Cart
- * @var $showWithTax
- * @var $suffix
  */
 ?>
 <ul id="mobile" class="list-group">
@@ -10,9 +8,7 @@
         <?php \Jigoshop\Helper\Render::output('shop/cart/mobile/'.$item->getType(), [
             'cart' => $cart,
             'key' => $key,
-            'item' => $item,
-            'showWithTax' => $showWithTax,
-            'suffix' => $suffix,
+            'item' => $item
         ]); ?>
     <?php endforeach; ?>
 </ul>