+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 ## Changelog
+* 2.1.4 - 2017.06.13:
+    * New: Email action for order status change for admin.
+    * New: Possibility to show prices with and without tax.
+    * New: All Forms now has data parameter.
+    * New: Shipping and Payment interfaces.
+    * Fix: Do not allow to add to cart variations with not enough stock.
+    * Fix: Show only selected attribute options for specified variable product.
+    * Fix: Api responses now properly includes result counts, next and prev paths, created/updated objects.
+    * Fix: PayPal error when order amount was more than 999.
+    * Fix: E-mail footer not included in Jigoshop emails.
+    * Fix: Non-existent tax classes supplied to TaxService.
 * 2.1.3 - 2017.05.25:
     * New: Allow to prepend product permalink with Wordpress permalink.
     * New: Tool to fix order items migration.